hello

+++ database/migrations/2025_02_02_003527_create_candidate_languages_table.php
@@ -5,16 +5,20 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration {
     public function up() {
-        Schema::create('candidate_languages', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('language_id'); // Doit être du même type que `languages.id`
-            $table->foreign('language_id')
-                ->references('id')
-                ->on('languages')
-                ->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('candidate_languages')) {
+            Schema::create('candidate_languages', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('language_id');
+                $table->foreign('language_id')
+                      ->references('id')
+                      ->on('languages')
+                      ->onDelete('cascade');
+                $table->timestamps();
+            });
+            
+        }
     }
+    
 
     public function down() {
         Schema::dropIfExists('candidate_languages');
